+++ builds.yml
@@ -19,6 +19,7 @@ buildTypes:
       - snaps
       - keyring-snaps
       - blockaid
+      - conf-redesign
     # Additional env variables that are specific to this build
     env:
       - INFURA_PROD_PROJECT_ID
